+++ package.json
@@ -45,7 +45,6 @@
     "markdown-to-jsx": "^7.3.2",
     "next": "^14.2.5",
     "next-i18n-router": "^4.1.1",
-    "next-yaml": "^1.0.1",
     "postcss": "8.4.36",
     "process": "^0.11.10",
     "publicodes": "1.1.1",
